% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(x)
}
\arguments{
\item{x}{: numeric vector, whose entries should be strictly between 0 and 1}
}
\value{
logit(x)
}
\description{
Takes in a vector and applies the logit function elementwise to that vector
}
