% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.r
\name{loglikelihood}
\alias{loglikelihood}
\title{loglikelihood}
\usage{
loglikelihood(M, I1, S1, I2, S2, l.iI, l.iS, logit.psi1, logit.psi2, alpha, s1,
  s2, s, mu, delta)
}
\arguments{
\item{M}{Number of replicates for the current exon. Positive integer.}

\item{I1}{Exon inclusion counts for group 1. Positive integers.}

\item{S1}{Exon skipping counts for group 1. Positive integers.}

\item{I2}{Exon inclusion counts for group 2. Positive integers.}

\item{S2}{Exon skipping counts for group 2. Positive integers.}

\item{l.iI}{Effective length of inclusion isoform. Positive integer.}

\item{l.iS}{Effective length of skipping isoform. Positive integer.}

\item{logit.psi1}{Numeric vector with values of logit psi1.}

\item{logit.psi2}{Numeric vector with values of logit psi2.}

\item{alpha}{Numeric vector with values of alpha.}

\item{s1}{Group 1 standard deviation. Positive number.}

\item{s2}{Group 2 standard deviation. Positive number.}

\item{s}{Overall standard deviation. Positive number.}

\item{mu}{Parameter mu.}

\item{delta}{Parameter delta.}
}
\value{
log likelihood value at input.
}
\description{
Used internally in PAIRADISE to compute the log-likelihood function
}
