% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize1.R
\name{optimize1}
\alias{optimize1}
\title{optimize1}
\usage{
optimize1(x, M, I1, S1, I2, S2, l.iI, l.iS, logit.psi1, logit.psi2, alpha,
  equal.variance = FALSE)
}
\arguments{
\item{x}{Numeric vector such that x = (sigma1, sigma2, sigma, mu, delta) if equal.variance = FALSE, and x = (sigma1, sigma, mu, delta) if equal.variance = TRUE. sigma1, sigma2, sigma must be positive}

\item{M}{Number of replicates for the current exon.}

\item{I1}{Exon inclusion counts for group 1. Positive integers.}

\item{S1}{Exon skipping counts for group 1. Positive integers.}

\item{I2}{Exon inclusion counts for group 2. Positive integers.}

\item{S2}{Exon skipping counts for group 2. Positive integers.}

\item{l.iI}{Effective length of inclusion isoform. Positive integer.}

\item{l.iS}{Effective length of skipping isoform. Positive integer.}

\item{logit.psi1}{Numeric vector with values of logit psi1.}

\item{logit.psi2}{Numeric vector with values of logit psi2.}

\item{alpha}{Numeric vector with values of alpha.}

\item{equal.variance}{Are the group variances assumed equal? Default value is FALSE.}
}
\description{
Used internally in PAIRADISE to compute the MLEs of delta, mu, sigma1, sigma2, sigma
}
