% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize2.r
\name{optimize2}
\alias{optimize2}
\title{optimize2}
\usage{
optimize2(x, k, I1, S1, I2, S2, l.iI, l.iS, delta, mu, s1, s2, s)
}
\arguments{
\item{x}{Numeric vector such that x = (logit(psi1), logit(psi2), alpha)}

\item{k}{Index representing current replicate number.}

\item{I1}{Exon inclusion counts for group 1. Positive integers.}

\item{S1}{Exon skipping counts for group 1. Positive integers.}

\item{I2}{Exon inclusion counts for group 2. Positive integers.}

\item{S2}{Exon skipping counts for group 2. Positive integers.}

\item{l.iI}{Effective length of inclusion isoform. Positive integer.}

\item{l.iS}{Effective length of skipping isoform. Positive integer.}

\item{delta}{Parameter delta.}

\item{mu}{Parameter mu.}

\item{s1}{Group 1 standard deviation. Positive number.}

\item{s2}{Group 2 standard deviation. Positive number.}

\item{s}{Overall standard deviation. Positive number.}
}
\description{
Used internally in PAIRADISE to compute the MLEs of logit(psi1), logit(psi2), alpha
}
