% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{: numeric vector}
}
\value{
sigmoid(x)
}
\description{
Takes in a vector and applies the sigmoid function elementwise to that vector
}
